chore: move Quest to plugins
